libui: Add hasSignaled() method to Fence
Adds a hasSignaled() method to android::Fence. This will make it easier
to write client code that only cares whether a fence has signaled, and
allows such code to gain the performance benefits of wait(0) (which is
significantly faster than getSignalTime()) while still being obvious
about what it is trying to do.

Test: m

Change-Id: Ia28ce6cff3860a171254ea20a24f3f1bfbc67eb8